BUG: Replace `exec_command` in `CCompiler.spawn`
`exec_command` is not thread safe. It uses os.spawnve which is not
thread safe on Windows (python issue #5309). However CCompiler.spawn
is used in a threaded environment (--parallel/-j build options).
This can lead to randomly segfaulting builds.

Replace `exec_command` with `subprocess.check_output`. Note that this
means that the command is no longer executed through the shell on posix
systems (hopefully no one relied on this behaviour).

Also remove arg quoting in gen_{preprocess,lib}_options as it is no
longer needed (and is in fact detrimental).